+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5445156_sys_FRESH-273_Migrate_C_OrderTax_set_ispackagingtax.sql
@@ -1,0 +1,3 @@
+update c_ordertax  
+set ispackagingtax = 'Y'
+where c_tax_id in (select c_tax_id from c_tax where name ilike '%Gebinde%');